+++ src/main.rs
@@ -6,6 +6,7 @@ pub mod common;
 pub mod consts;
 mod deploy;
 mod download;
+mod sign_as;
 pub mod socialdb_types;
 
 /// near-cli is a toolbox for interacting with NEAR protocol
